split test cases and changed the url in GetScript
