+++ README.md
@@ -7,4 +7,21 @@ A curated library of Redis Lua scripts
 
 ### Namespace
  - The first number of the lua script filename indicates the number of KEYS.
- Example: ```2_hsetxp.lua``` takes 2 keys.
+ Example: ```2_HSETXP.lua``` takes 2 keys.
+
+### Examples
+
+#### HASHES_XP
+Group ```HASHES_XP``` is HASHES with EXPIREAT by each field.
+Although you can set ```EXPIRE``` to ```HASHES```, hash set will expire by key so that you’ll discard all fields associated with the key as well.
+To avoid that, ```HASHES_XP``` creates two hash sets, one of those is normal hash set containing the key and fields, the other one hash set has time of the field expire at, whose key of the hash set is “<key>.EXPIREAT”.
+```go
+// To load lua script, execute GetScript
+script, err := GetScript("HASHES_XP/2_HSETXP.lua")
+
+// HSETXP key expire(secs) field value
+_, err := redis.Bool(script.Do(conn, "key", "10", "field", "value"))
+
+// check it works! b is true
+b, err := redis.Bool(conn.Do("HEXISTS", "key", "field"))
+```
