+++ lua/hashes_xp/1_hvalsxp.lua
@@ -5,7 +5,7 @@
 local HSET_EXPIREAT_KEY = KEYS[1]..".EXPIREAT"
 local now = tonumber(redis.call('TIME')[1])
 
--- get the list of fields in arguments, also values
+-- get the list of values in arguments
 local values = {}
 local k = ""
 local skip = false
