+++ lua/hashes_xp/2_hsetxp.lua
@@ -5,7 +5,6 @@
 local HSET_EXPIREAT_KEY = KEYS[1]..".EXPIREAT"
 local expireAt = tonumber(redis.call('TIME')[1]) + tonumber(KEYS[2])
 
--- get the list of fields in arguments, also values
 local k = ""
 for i, v in pairs(ARGV) do
     if i % 2 == 1 then
